Signed to unsigned

+++ lib/Modules/keypad.h
@@ -10,12 +10,12 @@ class KeypadHandler{
 class KeypadSolution{
     public:
     KeypadSolution();
-    void generate_solution(int seed);
+    void generate_solution(uint8_t seed);
 };
 
 class Keypad{
     public:
-    Keypad(int seed);
+    Keypad(uint8_t seed);
     void setup_module();
     void setup_gpio();
     void check_module();
